requirements.txt
torch
torchvision
torchdiffeq
numpy
scipy
scikit-learn
matplotlib
tqdm
PyYAML
wandb